# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audit/batch-pdf:
    post:
      operationId: createBatchAuditPdf
      tags: [audit]
      summary: Download combined multi-page accessibility report as PDF
      description: |
        Accepts an array of auditIds (from a prior batch scan) and generates a branded
        multi-page PDF: a cover page with a site-wide summary followed by a per-page
        section for each successfully scanned URL. Only include auditIds for successful
        pages (failed scans have an empty auditId and are not persisted).
        The cover page site score uses the same `totalChecks`-weighted formula as the batch SSE `complete` event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchAuditPdfBody"
      responses:
        "200":
          description: Multi-page PDF report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request (no auditIds provided or bad format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No audit results found for the provided IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: PDF generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CreateBatchAuditPdfBody:
      type: object
      properties:
        auditIds:
          type: array
          items:
            type: string
          minItems: 1
          description: Array of auditIds from successful batch page scans to include in the PDF
      required:
        - auditIds
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message